tunnelproxy, klient: remove debug cruft
